requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
flask-cors==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
gunicorn==21.2.0
